Basic functionality
Drones move in random patterns, but they share maps and eventually converge on a coordinate system. Proof of concept for the communication system.

Future work includes the ability to give them a pattern to try to match, and randomly adding/deleting bots to show the adaptibility of the network.

Mobile environments/environment hazards possibly after that.