Full-Featured Swarming
You can now draw a pattern, and the drones will swarm to it. They're not particularly smart about it, but they generally do a pretty good job of covering the pattern.

Future work:
Dynamically adding/removing drones.
Better pathing and drones to "push" others forward that have claimed a space but are blocking access to a further in space by telling that drone to take the space they're currently trying to reach.